Grand value division: value::ser
